requirements.txt
# Core AI/ML Libraries
google-generativeai==0.7.2
langchain==0.2.16
langgraph==0.2.50
langchain-google-genai==1.0.10

# Vector Store & Document Processing
llama-index==0.10.68
chromadb==0.5.20
sentence-transformers==3.3.1

# Web Scraping & API Integration
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.26.1
aiohttp==3.11.11
lxml==5.3.0

# Data Processing
pandas==2.2.3
numpy==1.26.4
pydantic==2.10.5
pydantic-settings==2.6.1

# Task Scheduling & Background Jobs
apscheduler==3.10.4

# Web Interface
streamlit==1.40.2
streamlit-chat==0.1.1

# Utilities
python-dotenv==1.0.1
loguru==0.7.3
rich==13.9.4

# Document Processing for Custom Sources
PyPDF2==3.0.1
python-docx==1.1.2

# Security & Production
cryptography==42.0.8
bleach==6.1.0
psutil==6.0.0

# Development
pytest==8.3.4
black==24.10.0 